+++ src/WebsiteInteractionSteps.php
@@ -350,4 +350,71 @@ JS
         }
     }
 
+    /**
+     * Opens specified page in default time
+     * Example: Given I am on "http://batman.com" in time
+     *
+     * @Given /^I am on "(?P<page>[^"]+)" in time/
+     * @Given /^Ich öffne die Seite "(?P<page>[^"]+)" innerhalb der Standardzeit$/
+     */
+    public function openUrl($page)
+    {
+        $this->visit($page);
+        $this->assertPageAddressWithinDefaultTimeout($page);
+    }
+
+    /**
+     * Opens specified page in specified time
+     * Example: Given I am on "http://batman.com" within 5 seconds
+     *
+     * @Given /^I am on "(?P<page>[^"]+)" within (?P<seconds>(\d+)) seconds?$/
+     * @Given /^Ich öffne die Seite "(?P<page>[^"]+)" innerhalb von (?P<seconds>(\d+)) Sekunden?$/
+     */
+    public function openUrlWithinSpecifiedTime($page, $seconds)
+    {
+        $this->visit($page);
+        $this->assertPageAddressWithinSpecifiedTime($page, $seconds);
+    }
+
+    /**
+     * Check if the current time is within the specified time. Otherwise, throw a PendingException and thus skip the test case.
+     * Example: When the current time is between "06:00" and "20:00", otherwise skip the test case
+     *
+     * @When /^the current time is between "(?P<fromTime>[^"]+)" and "(?P<toTime>[^"]+)", otherwise skip the test case/
+     * @When /^die aktuelle Uhrzeit liegt zwischen "(?P<fromTime>[^"]+)" und "(?P<toTime>[^"]+)", sonst breche das Testzsenario ab/
+     * @throws PendingException
+     */
+    public function actualTimeIsInSpecifiedTime($fromTime, $toTime){
+        date_default_timezone_set("Europe/Berlin");
+        $timestamp = time();
+        $currentTime = date("H:i",$timestamp);
+        $fromDate = date("H:i", strtotime($fromTime));
+        $toDate = date("H:i", strtotime($toTime));
+
+        if($this->checkTime($currentTime,$fromDate, $toDate) == false){
+            throw new \Behat\Behat\Tester\Exception\PendingException("The current time (" . $currentTime . ") is outside of the specified range. Specified range: from " . $fromDate . " to " .$toDate);
+        }
+    }
+
+    /**
+     * Check if the current time is within the specified time and return true or false
+     * Example: currentTime: 22:34 - fromeDate: 20:00 - toDate: 06:00 -> Result: false
+     */
+    private function checkTime($currentTime, $fromDate, $toDate){
+
+        if($fromDate < $toDate){
+            if(($currentTime >= $fromDate) && ($currentTime <= $toDate)) {
+                return true;
+            }else {
+                return false;
+            }
+        }elseif ($fromDate > $toDate){
+            if(($currentTime >= $fromDate) || ($currentTime <= $toDate)) {
+                return true;
+            }else {
+                return false;
+            }
+        }
+    }
+
 }